Some refactors on filter rules
